Bluetooth: Shell: Split GATT command to its own file
This makes it simpler to organize the GATT command implementation.
