+++ subsys/bluetooth/shell/Makefile
@@ -1,1 +1,2 @@
 obj-y += bt.o
+obj-y += gatt.o
