+++ subsys/bluetooth/shell/bt.h
@@ -1,0 +1,18 @@
+/** @file
+ @brief Bluetooth shell functions
+
+ This is not to be included by the application.
+ */
+
+/*
+ * Copyright (c) 2017 Intel Corporation
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#ifndef __BT_H
+#define __bt_H
+
+extern struct bt_conn *default_conn;
+
+#endif /* __BT_H */
