+++ subsys/bluetooth/shell/gatt.h
@@ -1,0 +1,29 @@
+/** @file
+ @brief GATT shell functions
+
+ This is not to be included by the application.
+ */
+
+/*
+ * Copyright (c) 2017 Intel Corporation
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#ifndef __GATT_H
+#define __GATT_H
+
+int cmd_gatt_exchange_mtu(int argc, char *argv[]);
+int cmd_gatt_discover(int argc, char *argv[]);
+int cmd_gatt_read(int argc, char *argv[]);
+int cmd_gatt_mread(int argc, char *argv[]);
+int cmd_gatt_write(int argc, char *argv[]);
+int cmd_gatt_write_without_rsp(int argc, char *argv[]);
+int cmd_gatt_write_without_rsp_repeated(int argc, char *argv[]);
+int cmd_gatt_write_signed(int argc, char *argv[]);
+int cmd_gatt_subscribe(int argc, char *argv[]);
+int cmd_gatt_unsubscribe(int argc, char *argv[]);
+int cmd_gatt_register_test_svc(int argc, char *argv[]);
+int cmd_gatt_write_cmd_metrics(int argc, char *argv[]);
+
+#endif /* __GATT_H */
